+++ src/test/suite/ALExtractProcedureCA.test.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
 import * as vscode from 'vscode';
-import { ALExtractToProcedureCA } from '../../extension/Code Actions/alExtractToProcedureCA';
+import { ALExtractToProcedureCA } from '../../extension/Services/alExtractToProcedureCA';
 import { ALProcedure } from '../../extension/Entities/alProcedure';
 import { ReturnTypeAnalyzer } from '../../extension/Extract Procedure/returnTypeAnalyzer';
 import { ALLanguageExtension } from '../alExtension';
